Move to common shared steps for data and metadata
