+++ src/components/rules-creation-page/common/steps/GeneralInfoStep.jsx
